Refactor modal handling: implement custom hook for modal state management and integrate it across trip management components.
