+++ .vscode/settings.json
@@ -29,7 +29,7 @@
     "Tripzy"
   ],
   "github.copilot.enable": {
-    "*": true,
+    "*": false,
     "plaintext": false,
     "markdown": false,
     "scminput": false,
