+++ README.md
@@ -39,7 +39,7 @@ Instalar con npm
 [ ] En members agregar el back y funcionalidad para invitar
 [ ] Proteger ruta de trip si el usuario no esta invitado
 [ ] Mostrar los viajes que pertenecen al usuario en profile (los admin deberían ver todos)
-[ ] Hacer un custom hook para el modal (y cambiar donde se usa actualmente)
+[ x ] Hacer un custom hook para el modal (y cambiar donde se usa actualmente)
 [ ] Proteger api por rol
 [ ] Agregar un boton en TripLayout para poder bloquear el viaje si sos admin
